@@ -40,27 +40,29 @@
     "@types/memdown": "^3.0.0",
     "@types/node-localstorage": "^1.3.0",
     "@types/redux-logger": "^3.0.7",
-    "@typescript-eslint/eslint-plugin": "^1.5.0",
-    "@typescript-eslint/parser": "^1.5.0",
-    "eslint": "^5.15.3",
+    "@types/tiny-async-pool": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^1.6.0",
+    "@typescript-eslint/parser": "^1.6.0",
+    "eslint": "^5.16.0",
     "eslint-config-prettier": "^4.1.0",
     "eslint-plugin-prettier": "^3.0.1",
-    "ganache-cli": "^6.4.1",
-    "jest": "^24.5.0",
+    "ganache-cli": "^6.4.2",
+    "jest": "^24.7.1",
     "jest-junit": "^6.3.0",
     "memdown": "^3.0.0",
     "node-localstorage": "^1.3.1",
     "prettier": "^1.16.4",
     "rimraf": "^2.6.3",
     "rxjs-marbles": "^5.0.0",
-    "ts-jest": "^24.0.0",
+    "tiny-async-pool": "^1.0.4",
+    "ts-jest": "^24.0.2",
     "ts-node": "^8.0.3",
     "typechain": "^0.3.14",
-    "typescript": "^3.3.4000"
+    "typescript": "^3.4.2"
   },
   "dependencies": {
     "ethers": "^4.0.27",
-    "io-ts": "^1.8.4",
+    "io-ts": "^1.8.5",
     "lodash": "^4.17.11",
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
@@ -85,7 +87,9 @@
       "ecmaVersion": 2018,
       "sourceType": "module"
     },
-    "rules": {},
+    "rules": {
+      "@typescript-eslint/no-non-null-assertion": "off"
+    },
     "extends": [
       "plugin:@typescript-eslint/recommended",
       "prettier/@typescript-eslint",
